# [Enhancement] use cascade merge as exchange merge