--- a/be/src/runtime/data_stream_recvr.h
+++ b/be/src/runtime/data_stream_recvr.h
@@ -38,7 +38,8 @@ namespace starrocks {
 
 namespace vectorized {
 class SortedChunksMerger;
-}
+class CascadeChunkMerger;
+} // namespace vectorized
 
 class DataStreamMgr;
 class MemTracker;
@@ -172,6 +173,7 @@ private:
 
     // vectorized::SortedChunksMerger merges chunks from different senders.
     std::unique_ptr<vectorized::SortedChunksMerger> _chunks_merger;
+    std::unique_ptr<vectorized::CascadeChunkMerger> _cascade_merger;
 
     // Pool of sender queues.
     ObjectPool _sender_queue_pool;