--- a/be/src/runtime/sorted_chunks_merger.cpp
+++ b/be/src/runtime/sorted_chunks_merger.cpp
@@ -4,6 +4,9 @@
 
 #include "column/chunk.h"
 #include "exec/sort_exec_exprs.h"
+#include "exec/vectorized/sorting/sorting.h"
+#include "runtime/chunk_cursor.h"
+#include "runtime/runtime_state.h"
 
 namespace starrocks::vectorized {
 
@@ -95,7 +98,7 @@ void SortedChunksMerger::init_for_min_heap() {
         _min_heap.reserve(_cursors.size());
         for (auto& cursor_ptr : _cursors) {
             ChunkCursor* cursor = cursor_ptr.get();
-            cursor->reset_with_next_chunk_for_pipeline();
+            cursor->next_chunk_for_pipeline();
             cursor->next_for_pipeline();
             if (cursor->is_valid()) {
                 _min_heap.push_back(cursor);
@@ -284,6 +287,42 @@ void SortedChunksMerger::collect_merged_chunks(ChunkPtr* chunk) {
     _result_chunk->set_num_rows(_row_number); // set constant column in chunk with right size.
     (*chunk) = std::move(_result_chunk);
     _row_number = 0;
+}
+
+CascadeChunkMerger::CascadeChunkMerger(RuntimeState* state, RuntimeProfile* profile)
+        : _state(state), _profile(profile), _sort_exprs(nullptr) {}
+
+Status CascadeChunkMerger::init(const std::vector<ChunkProvider>& providers,
+                                const std::vector<ExprContext*>* sort_exprs, const std::vector<bool>* sort_orders,
+                                const std::vector<bool>* null_firsts) {
+    std::vector<std::unique_ptr<SimpleChunkSortCursor>> cursors;
+    for (int i = 0; i < providers.size(); i++) {
+        cursors.push_back(std::make_unique<SimpleChunkSortCursor>(providers[i], sort_exprs));
+    }
+    _sort_exprs = sort_exprs;
+    _sort_desc = SortDescs(*sort_orders, *null_firsts);
+
+    _merger = std::make_unique<MergeCursorsCascade>();
+    RETURN_IF_ERROR(_merger->init(_sort_desc, std::move(cursors)));
+    return Status::OK();
+}
+
+bool CascadeChunkMerger::is_data_ready() {
+    return _merger->is_data_ready();
+}
+
+Status CascadeChunkMerger::get_next(ChunkPtr* output, std::atomic<bool>* eos, bool* should_exit) {
+    if (_merger->is_eos()) {
+        *eos = true;
+        return Status::OK();
+    }
+    ChunkUniquePtr chunk = _merger->try_get_next();
+    if (!chunk) {
+        *should_exit = true;
+        return Status::OK();
+    }
+    *output = ChunkPtr(chunk.release());
+    return Status::OK();
 }
 
 } // namespace starrocks::vectorized