@dulong41 2022-11-08
run starrocks_admit_test