@mofeiatwork 2022-11-09
run starrocks_admit_test