@mofeiatwork 2022-11-09
fix starrocks_admit_test