@mofeiatwork 2022-11-10
run starrocks_admit_test