@wanpengfei-git 2022-11-11
run starrocks_admit_test